+++ source/index.html
@@ -451,10 +451,7 @@
       <div class="screen__wrapper">
         <div class="result result--negative">
           <div class="result__image">
-            <picture>
-              <source srcset="img/result3-mob.png 1x, img/result3-mob@2x.png 2x" media="(max-width: 768px) and (orientation: portrait)">
-              <img src="img/result3.png" srcset="img/result3@2x.png 2x" alt="">
-            </picture>
+            <canvas class="result_canvas" id="crocodile-scene"></canvas>
           </div>
           <h2 class="result__title">
             <span class="visually-hidden">Не угадал!</span>
